incorporate unplugin-icons as a preparation to migrate to Vue3
BootstrapVueNext does not support icons, so unplugin-icons was adopted
to replace the BootstrapVue icon components as part of the migration
to Vue 3.

Additionally, unplugin-vue-components/vite supports tree shaking, so
unused icons are automatically removed. With the configuration in
vite.config.js, icons are automatically imported without needing to be
manually included in individual components.